+++ packages/react/src/client/types/ToolUI.ts
@@ -1,14 +1,15 @@
 import { Unsubscribe } from "@assistant-ui/tap";
 import { ToolCallMessagePartComponent } from "../../types";
 
-export type AssistantToolUIState = Record<
-  string,
-  ToolCallMessagePartComponent[]
->;
+export type ToolUIState = Record<string, ToolCallMessagePartComponent[]>;
 
-export type AssistantToolUIActions = {
+export type ToolUIActions = {
   setToolUI(
     toolName: string,
     render: ToolCallMessagePartComponent,
   ): Unsubscribe;
+};
+export type ToolUIMeta = {
+  source: "root";
+  query: Record<string, never>;
 };
